+++ src/SFA.DAS.Reservations.Application/Employers/Queries/GetTrustedEmployersQuery.cs
@@ -4,6 +4,6 @@ namespace SFA.DAS.Reservations.Application.Employers.Queries
 {
     public class GetTrustedEmployersQuery : IRequest<GetTrustedEmployersResponse>
     {
-        public long UkPrn { get; set; }
+        public uint UkPrn { get; set; }
     }
 }
